requirements.txt
selenium
lxml
requests
mysql-connector-python
google-api-python-client
pyvirtualdisplay
sqlalchemy
pandas
google_auth_oauthlib
google_auth
pyOpenSSL
oauth2client
sendgrid
colorama
dnslib
tldextract
